python/requirements.txt
fastapi>=0.112.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
